% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalDeliverable.R
\docType{data}
\name{CF_Biomarkers}
\alias{CF_Biomarkers}
\title{CF Data Set from Positive and Negative CF patients}
\format{A data frame with 216 rows and 4 variables:
\describe{
  \item{m/z}{Mass to Charge ratio}
  \item{RMT}{Relative Migration Time}
  \item{Sample}{The patient of which the sample was take/ patient number}
  \item{RPA}{Relative Peak Area}
}}
\usage{
data(CF_Biomarkers)
}
\description{
A dataset containing information about the different relative peak areas for biomarker peaks identified with specific Relative Migration Times and mass to charge ratios
}
\keyword{datasets}
