% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalDeliverable.R
\docType{data}
\name{CF_BiomarkersPCA}
\alias{CF_BiomarkersPCA}
\title{CF Data Set from Positive and Negative CF patients displayed with patients as variables}
\format{A data frame with 49 rows and 25 variables:
\describe{
  \item{Label}{the mass/z ratio : the Relative migration Time}
  \item{CF_1}{RPAs for CF patient 1}
  \item{CF_2}{RPAs for CF patient 2}
  \item{CF_3}{RPAs for CF patient 3}
  \item{CF_4}{RPAs for CF patient 4}
  \item{CF_5}{RPAs for CF patient 5}
  \item{CF_6}{RPAs for CF patient 6}
  \item{CF_7}{RPAs for CF patient 7}
  \item{CF_8}{RPAs for CF patient 8}
  \item{CF_9}{RPAs for CF patient 9}
  \item{CF_10}{RPAs for CF patient 10}
  \item{CF_11}{RPAs for CF patient 11}
  \item{CF_12}{RPAs for CF patient 12}
  \item{NEG_1}{RPAs for CF patient 13}
  \item{NEG_2}{RPAs for CF patient 14}
  \item{NEG_3}{RPAs for CF patient 15}
  \item{NEG_4}{RPAs for CF patient 16}
  \item{NEG_5}{RPAs for CF patient 17}
  \item{NEG_6}{RPAs for CF patient 18}
  \item{NEG_7}{RPAs for CF patient 19}
  \item{NEG_8}{RPAs for CF patient 20}
  \item{NEG_9}{RPAs for CF patient 21}
  \item{NEG_10}{RPAs for CF patient 22}
  \item{NEG_11}{RPAs for CF patient 23}
  \item{NEG_12}{RPAs for CF patient 24}
}}
\usage{
data(CF_BiomarkersPCA)
}
\description{
A dataset containing information about the different RMT for biomarker peaks identified with specific Relative Migration Times and mass to charge ratios
}
\keyword{datasets}
