% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalDeliverable.R
\docType{package}
\name{FinalDeliverable}
\alias{FinalDeliverable}
\title{FinalDeliverable: A package with a minimum package creation}
\description{
FinalDeliverable: A package with a minimum package creation
}
\references{
\url{https://github.com/paezha/Reproducible-Research-Workflow}
}
\author{
Megan Magee, Department of Chemistry and Chemical Biology, McMaster University \email{mageem@mcmaster.ca}
}
